ai-empathy-test/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
streamlit
python-dotenv
transformers
torch  # Required by transformers